safe-therapy-agent/requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
typing-extensions>=4.8.0







